+++ src/main/kotlin/io/korti/adventofcode/intcode/Parameter.kt
@@ -1,3 +1,3 @@
 package io.korti.adventofcode.intcode
 
-data class Parameter(val mode: ParameterMode = ParameterMode.POSITION_MODE, val code: Int)
+data class Parameter(val mode: ParameterMode = ParameterMode.POSITION_MODE, val code: Long)
